[OPLU-37 JSON format validation (#27)
Adds file-related validators. There are no validations for a text file. For JSON files, we check:
if the file is a valid JSON,
if the file has required keys (witnesses, table),
if there is at least one witness,
if the number of tokens in each row is the same as the number of witnesses,
I also added that the import errors would be saved on Project.

Co-authored-by: Agnieszka Pastuszczak <agnieszka.pastuszczak@xfive.co>