[OPLU-250] Allow selecting readings for tokens created from selected_text (#87)
Adds a flag to tokens -- resized. Every token that is created from the selected text will have it set to true. Other tokens (for the text before/after selection) won't have it. 

Updates the algorithm for calculating the state. If the token has resized: true, then the number of grouped variants/readings does not matter and the token is displayed as the one that needs a selection.